replace integer field to int in second

+++ first/models.py
@@ -5,5 +5,5 @@ from django.db import models
 class History(models.Model):
     date = models.DateTimeField(default=0)
     first = models.IntegerField(default=0)
-    second = models.IntegerField(default=0)
+    second = models.FloatField(default=0)
     result = models.IntegerField(default=0)
